Update business_owner schema to prepare for KYC-ed BOs (#2957)
## Describe your changes

The owner that creates the business will be designated primary in the DB
and have a user_vault_id attached to it immediately. The link_id for now
doesn't matter, but it would be nice if this matched the link_id inside
the JSON blob in the vault...

The other BOs that are invited will be designated secondary and will not
start with a user_vault_id. But, they will have a link_id that matches
one of the BOs inside the JSON blob in the vault