[ci] Use more produciton-like bifrost client (#2972)
## Describe your changes

as we're about to write very complex tests that run multi-KYC
integration tests for KYB, this is an overhaul of the BifrostClient that
acts more like real bifrost - we actually hit the `GET
/hosted/onboarding/status` endpoint, look at the requirements, and
complete them in the same way that real bifrost does. the ob config we
feed into the bifrost client now wholly defines what information will be
provided to complete the user's onboarding

used this as a way to audit use of BifrostClient as well. now, we
instantiate the BifrostClient with a set of default data that can be
overriden before `.run()`ing bifrost and actually sending the data to
the backend. the User returned from `.run()` also contains a reference
to the client that created it, so you can always find what info was sent
to the backend

## Caveats

Still some cleanup that can be done after this 

## Test Plan

Tests :)