[FP-3457] Create a BoSession auth token (#2977)
## Describe your changes

Creates a new form of session-based auth that we will send out to
secondary BOs in order to complete KYC.

This token both uniquely identifies an ob config (which is why it is
considered a form of ob config auth) AND uniquely identifies a specific
BusinessOwner row that is linked to a business vault but doesn't yet
have a KYCed user vault.

This token allows us to not have to pass the ob config public key when
the BO fills out their KYC form. We will also use this token to display
some information about the business and who invited this individual to
join the business

This auth token will be acceptable in anything where we currently take
`tenantPk`: the identify flow and `GET /org/onboarding_config`

Also resolves FP-3458

## Test Plan

Will test in future PR