[FP-3459] Link user vault to business when BoSessionAuth is provided in identify (#2989)
## Describe your changes

In the identify flow, we accept (1) a tenant PK (2) a short-lived
onboarding session token for security-conscious tenants and now (3) a
KYB token

When a KYB token is provided, we know that we are creating/locating a
user for the purpose of KYCing them as a secondary BO. So, we can update
the BO row with the user vault ID that was located.

Tangentially, in `POST /hosted/user/onboarding`, we create a new
business for a user if the ob config requires business data. Added a
codepath that doesn't do this if there's already a business associated
with the auth token (which is the case for these secondary BOs' tokens)

Also resolves FP-3464

## Caveats

This token also technically gives the secondary BO the permissions to
update the business vault. Is this the behavior we want?

## Test Plan

Wrote a basic integration test that runs the secondary BO through KYC
using the kyb token instead of the tenant PK. The secondary BO row in
the DB now has the user_vault_id attached:
```
footprint_db=# select user_vault_id, business_vault_id, kind from business_owner order by _created_at desc;
       user_vault_id       |     business_vault_id     |   kind
---------------------------+---------------------------+-----------
 uv_QwmDHctJ9eWGFdzPEB3L53 | bv_CATpMXS84vpauGFDgGDMiV | secondary
 uv_dpTkF38PIkGy67CTJm3ZAI | bv_CATpMXS84vpauGFDgGDMiV | primary
```