[FP-3460] Add endpoint to get business information before KYCing secondary BO (#3006)
## Describe your changes

The hosted bifrost flow will hit this endpoint to get various info to
display on the business being verified and who invited them to verify
the business

Can add/rm info from this API as we start to integrate with it

## Caveats

Right now, we get the name of the person that invited you from the
business vault's KYCed BOs. The primary BO in that list is supposed to
have the same name as the vault for the primary BO, but we don't
currently enforce this anywhere on the backend - only on the frontend.
Maybe we will add a check for this on the backend, or maybe we will just
rely on the fronted

## Test Plan

Integration test!