[FP-3499] Have link ID work for primary BOs as well (#3007)
## Describe your changes

The code to zip BOs from the vault and BOs from the DB didn't work for
primary BOs. We generate the primary BO in the DB before the primary BO
in the vault, and it's really difficult to propagate the link_id of the
primary BO in the DB all the way into the vaulting code. So, we just
generate the same fixed ID in each place to solve this problem

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

Integration tests