[FP-3541] Add document fields with new structure (part 5) (#3060)
## Describe your changes

Enable id-docs again, using the new endpoints. Important to mention:
we're not retrieving the status of each upload yet, because we're making
some changes in how the backend is handling these data. Given we don't
have users using it yet, this shouldn't be a big problem for now. I'll
add some tests in a follow up PR


https://user-images.githubusercontent.com/446731/231849332-9cbc8968-9584-4f4a-9c31-9909a928fdda.mov


## Issue: FP-3541

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->