Generify gaurds to reuse for user auth (#3080)
## Describe your changes

A lot of the auth primitives that we created for tenant auth can be
reused for user auth - this does some refactoring of tenant auth to move
things into shared places. For ex, there's no reason to make a specific
`Or` struct or `CanDecrypt` struct for user auth and for tenant auth -
we're able to reuse them!