Add new decrypt API (#3082)
## Describe your changes

With our new user auth that looks similar to tenant auth, we can now use
the exact same CanDecrypt struct simply by implementing the
`IsMet<UserAuthScope>` trait for CanDecrypt!

This adds an actual decrypt API that we'll use inside of bifrost to get
information on file for a user. In the future, the auth method will
determine what scopes the user token has - biometric auth will give more
permissions.

Will write a test for this API shortly and refine the permissions