[FP-3569] Remove difference between OrgOnboarding and OrgOnboardingInit scopes (#3108)
## Describe your changes

Removes the difference between `OrgOnboarding` and `OrgOnboardingInit` auth scopes. This difference hasn't meant much for a long time (since we started serializing scoped vault ID in OrgOnboardingInit instead of OrgOnboarding), but user auth needed some TLC and i just haven't gotten around to it. 

Before this PR, a user token had OrgOnboardingInit immediately after the identify flow, but only the OrgOnboarding scope after calling `POST /hosted/onboarding`. The presence of the OrgOnboarding scope was functionally identical to whether there is an onboarding for the scoped user

## Caveats

Removing an auth scope is a backwards-incompatible change since existing auth tokens in the DB still need to be deserialized. This preserves the old `OrgOnboarding` scope in code but stopes creating them in the DB. we'll only remove it from code once we have no more active tokens left in the DB

## Test Plan

Existing integration tests