Change return type of POST /users (#3099)
## Describe your changes

We're about to get rid of the wire type we used to return here, so
switching to a more bare-bones serialization.

## Caveats

This is a tenant-facing API, which is normally hard to change. but it's
only used to create a new non-portable user when migrating old legacy
non-footprint users onto footprint. It's very unlikely anyone is
actively using this (we'd see non-portable vaults being created) and
relying upon the response

## Test Plan

<!--- How did you test your changes? :) -->