Add seqno to get_bulk_active_for_tenant (#3122)
## Describe your changes

Extends `DataLifetime::get_bulk_active_for_tenant` to take in an
optional seqno.

For Multi-KYC KYB, when all BO's have completed KYC and we want to kick
off the KYB call to Middesk, we populate BO data in the request from the
real vaults of those BO's, so we'd like to use this bulk retrieval
method there. Currently, when we populate a VaultWrapper from a VReq for
KYC vendor calls, we filter to the seqno of that vreq. I think its
reasonable to do the same for KYB so extending the bulk query to support
this

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->