[FP-3614] Add simple GET /users to search on fingerprints (#3139)
## Describe your changes

This is a totally separate API from our internal-only `GET /entities`
APIs that we use in the dashboard. We want our docs to be able to show
even a smidge of the capabilities we provide, like searching on
fingerprints. So, we need a very simple `GET /users` API that includes
the ability to search

## Test Plan

Integration test