Rm ob_config_id in onboarding completed event (#3213)
## Describe your changes

We also removed this from the `POST /validate` response because the
tenants don't keep track of ob configs themselves anyways. they could
theoretically go look up an ob config by the ID in this webhook, but i
can't see a use case for it yet. For now, going to rm until a tenant
needs it

## Caveats

Breaking change :o but i doubt anybody is using this