[FP-3656] Support a fixture phone number (#3212)
## Describe your changes

To help make testing in sandbox easier, we can offer a fixed phone
number that always has a fixed pin code.

Vaults made with this phone number will not be made portable across
tenants, which we ensure by not creating globally-scoped fingerprints

## Caveats

After building this, I'm not quite sure it's worth the risk - i
initially was bought in because shopify/stripe provide for ex fixture
card numbers, but this is a kind of dangerous codepath in a few ways:
- Could accidentally remove the code that doesn't create global
fingerprints
- Could accidentally remove the check that the fixed phone number only
works in sandbox and provide a backdoor to onboard in prod

But, probably shows good faith with customer for now. Can always
deprecate in the future. Lmk what you think

## Test Plan

Integration test for identify flow w fixed PIN.

I looked into migrating all of our integration tests to use this fixed
phone number, but it's a decent bit of surgery. If the twilio bill gets
annoying, we could do this since we short circuit and don't send
messages to this phone number