Send real link to hosted bifrost for multi-KYC KYB (#3245)
## Describe your changes

We are ready!!!

## Caveats

The link sent by the backend should really be a function of the
environment used on the frontend (vs the env on the backend as in this
PR today). For ex, if you are running the frontend locally pointed at
prod, you should get a text with a link to your localhost version of
hosted bifrost. For d2p, we do actually have the frontend generate the
link to send. This is kind of annoying in this case since links are sent
as a side effect of adding data to the vault.

I think we'll want to build a more general approach to this where the
frontend just always tells the backend what environment is running on
the frontend and the backend will use that info to generate the correct
link