[FP-3524] Send email link to BOs to KYC (#3246)
## Describe your changes

In addition to the SMS we send, also send an email.

## Caveats

Still some kinks to work out in the template - for ex, might want to
show the name of the tenant in the email or SMS copy. Kind of weird that
we just show the logo optionally of the tenant with no more context

## Test Plan

Sent email to self:


![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/1df2a0b7-a979-4921-82f7-b8ff60c76514/image.png)