Create a new appearance table (#3286)
## Describe your changes

For hosted bifrost, there's no way for a tenant to specify an appearance
without us storing it on the backend. We had a similar problem for d2p
(which is similarly hosted by us) and will have a similar problem for
embedded components.

This adds a very very primitive customized theme table called
`appearance` that is basically a json blob - each ob config can have its
own appearance, for now.

Eventually, we may have a full appearance configurer in the dashboard -
you could imagine the developers tab may have a list of appearances for
the tenant and each ob config can link to an existing appearance.

For now, we will just create these manually in the backend to unblock
hosted bifrost

## Caveats

very v0

## Test Plan

none, will test when we have an ob config with an appearance