[FP-3835] Migrate vault endpoints to PATCH alongside PUT (#3304)
## Describe your changes

I don't really know why i initially made these a `PUT`.... in REST, a
PUT operation is meant to entirely replace an object with the one
provided (thus overwriting any data not provided with null) while a
PATCH operation is meant to overlay a subset of fields on the object
provided on top of the existing object. Maybe when we had fewer kinds of
data it felt more like a PUT that fully replaced the vault. not sure why
i made it PUT....

Just want to migrate this before we have lots of people looking at our
APIs. One knowledgable in REST may expect that the PUT operation would
unset any keys not provided in the request

## Caveats

This just adds the PATCH alongside the PUT (and marks the old PUT as
deprecated). I'll migrate all the hosted APIs since the only client is
us - for the tenant-facing APIs, i can likely deprecate the `PUT
/entities/<fp_id>/vault`, but might have to keep `PUT
/users/<fp_id>/vault`... not sure, i don't think anybody is really using
the `PUT /users/<>/vault` API so we could probably rename it as long as
there is an equivalent endpoint for them to use

## Test Plan

future PR