[FP-3802] Clean up GET user/token api (#3288)
## Describe your changes

Frontend will start using this to see if the auth token has expired. Just making one cleanup to allow this to be hit with any auth token

## Caveats

This does expose auth token scopes publically, but we have been doing that for a while. no real risk in it since security is not via obscurity