[FP-3815] Add OptionalJson<T> extractor to stop swallowing errors (#3298)
## Describe your changes

`Option<Json<T>>` would swallow any errors json-parsing T because it maps Err results to `None`.

This is a new extractor that only results in `None` if there is no body at all to be parsed

## Caveats

Two of these cannot be used in parallel on one extractor - oh well!

## Test Plan

Uncommented the integration test that was broken before this