Incode State machine (#3252)
## Describe your changes

A state machine that 
* initializes itself into a particular state.
For now, it always initializes to start an onboarding. Once we have an
incode specific state management table, we'll be able to re-use the
token and start from adding documents. This will be done in the next
branch

States:
* define the logic to execute that state, and then return a `Transition`
to the next state based on what happens

Since for document upload, we need to handle various errors differently
(e.g. some transient DB issue vs. a front image coming back blurry) so
we return back to the caller the state + error to appropriately handle.
This might change once work through the exact details of retrying, but
this feels like a good place to stop until I actually start integrating
into /documents

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

wrote an e2e test