Add new AssertionError util (#3325)
## Describe your changes

The current `ApiError::AssertionError` is tricky to use because it
doesn't accept `&str` (we'd have to parameterize all of ApiError with a
lifetime, which maybe we will do at some point). To make it worse, we
often will make an `ApiError::AssertionError` inside of an `.ok_or()`,
which clippy then complains needs to be `.ok_or_else(|| )` since we
stringify a static `&str` inside the lambda.

This just adds a little util that makes it easy to add assertion errors
with a simple `&'static str` message.

In general, i realized that i was adding some `BusinessError::blah`
custom error types for these assertion errors, but those would return
`HTTP 400`s to the client, leaving them confused what input they could
edit to get around this validation error. If there's a true application
invariant broken (like a portable user vault that has no phone number),
we should 500 so we know it's something for us to look into

## Caveats

Is definitely not the terminal state for error handling - there's a lot
more work to be done, but this is step in the right direction :)