Allow duplicate phones/emails in sandbox (#3374)
## Describe your changes

Implements the same logic as #3372 but on the frontend side to make
testing a little easier

## Caveats

This isn't my favorite solution... feels strange to add more branches on
sandbox. But it definitely helps to make testing easier - otherwise a
dev doing testing needs to send the BO link to a secondary phone and
email. lmk what you think :)

## Test Plan

Tested locally