[FP-3923] Start serializing an onboarding requirement for authorize (#3385)
## Describe your changes

We're migrating the authorize page to be its very own requirement in
order to allow us to one day add step up flows were requirements are
added after the authorize screen.

Now, we serialize a new requirement when the onboarding isn't
authorized. This isn't a breaking change because the frontend ignores
any requirements it doesn't know about

## Test plan

Tested bifrost locally, we proceed normally even with this new
requirement from the backend