return enum from incode state machine (#3395)
## Describe your changes

We want to be able to materialize information we compute/fetch from
incode during the state transitions.

It's also nice to be able to assert things about various states. If we
return a `Box<dyn Trait>` we lose easy access
to the concrete types that implement the state transition. By using
enum_dispatch, we can recover the types themselves (and also speed up
perf)

Also adds some initialization code that is required for impl'ing retries
in /document

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

updated existing tests