Perform additional validation when authing as a user (#3443)
## Describe your changes

As we're going to start having a handful of different paths that generate auth tokens, it's easier to break invariants. Let's start to enforce those invariants as soon as the token is used to auth.

For now, we just enforce that you don't create a user auth token for a non-portable vault or a business vault

Also kind of follows the more modern pattern where the `Session<T>` representing the auth extractor has a `T` that just includes various context that might be useful when processing the request - now we pre-fetch the `Vault` in the extractor to validate, so we also just store it on the session `T`.

## Caveats

Not making the code any cleaner - it's difficult to know what is stored serialized in the `session` table vs what is loaded and populated dynamically at runtime. auth code is looking for some sweeping up in the near future

## Test Plan

Should be backwards compatible