[FP-4081] Start decrypting populated info (#3459)
## Describe your changes

- In bifrost, the onboarding requirements machine now always renders the
`CollectKycData` plugin one time in order to show the confirm page.
- Note: there may be some times (ie document-only upload) where we don't
want to render the collect kyc data plugin at all, and in those cases,
the backend won't even return this as a met requirement
- In the CollectKycData plugin, we now decrypt basic attributes that are
already populated on the user vault
- We don't decrypt sensitive data, like ssn, because we don't know if
our auth method will yet allow it. Need to implement step-up in a
follow-up
- Add support for values that exist in the vault but aren't yet
decrypted
- We render these on the confirm page the same way as we do in the
dashboard as `•••••••••`
- But the edit pages for these values have an empty value, requiring the
user to specify a new value to edit it

## Caveats

To prevent this from affecting the demo of the one-click flow while the
feature is in development, i started hiding met requirements from the
backend for the demo tenant here: #3463 this will cause us to skip past
showing the confirm screen

This won't play super well with cases where you have a full ssn9 on file
and the new ob config requires only ssn4 - need to play around with that

The code to prevent showing the "additional info required" screen when
the KYC requirement is already met is a little fugly... i have an idea
on how to fix this, but will be a larger full-stack change that i'll
make tomorrow :)

## Test Plan

Lots of local testing... unit tests

Demo here:
- first, we onboard onto an ob config that requires only ssn4 
- then, we onboard onto a config requiring full ssn. we properly collect
the full ssn9
- we show the confirm screen with sensitive fields scrubbed. when
editing specific pieces of info, we are required to populate the
scrubbed info
- show in the network tab that we only vault newly collected data


https://github.com/onefootprint/monorepo/assets/21066341/21cb33f5-9854-4aa7-a978-abc370b24416

And, after all of this, we then onboard the user onto an ob config for
which everything is already collected. we skip straight to the confirm
screen!





https://github.com/onefootprint/monorepo/assets/21066341/c09bd1ad-8cae-45cf-8dec-26286a177d04