[FP-3970] Start saving some OCR data into the user vault (#3511)
## Describe your changes

When the incode machine finishes, save some OCR data into the vault at the same time as we save the documents

## Caveats

For now, i'm just using a custom DI to store this info. We could make this a strongly-typed metadata DI if we want, but good for POC now. Not sure if tenants will ever decrypt this or if we'll just use it when proxying to alpaca

I also realized that all of the API responses don't use proper `Pii` types - we'll want to make a library of more thorough `PiiInt`, `PiiXXX` types that we use in every single vendor client - otherwise this is an easy way to leak PII

## Test Plan

Integration tests using some fixture data