refactor get_fixture_data_decision (#3504)
## Describe your changes

`get_fixture_data_decision` currently requires a UVW because it needs to
get the decrypted phone number and check for fixtures defined via the
`#`. Right now various callsites are having to materialize a UVW and
pass it in. This just refactors this so the function creates the UVW
itself. The method already has to be `async` because of the enclave call
so its not a huge deal to do this db_pool op here too.

Convenience since I'll be calling this from a few places in workflow
code. Also discourages callsites from making a UVW for this method call
but then using it for other things when really a UVW should be
materialized in a txn for that thing, eg:
https://github.com/onefootprint/monorepo/pull/3504/files#diff-26944df0583fd7d66e2f943eff6f8db1686be059c1883a904bfb9f9ce38a037fR260

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->