Mock vendor calls for sandbox/demo (#3505)
## Describe your changes

Incorporates Sandbox logic into KYC workflow by mocking vendor calls if
sandbox. For now we just have 1 canned response for each possible
VendorAPI. In the future, we could pick one out of several fixture
responses depending on the Sandbox fixtures chosen but this is a little
nuanced in that outcomes like Stepup and Review are functions of rules
and not just the vendor data itself. (And one immediate hitch here is
our rules always raise review when kyc=Fail, but in our fixtures we
allow a user to chose either hard fail or fail with review)


## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->