[FP-4176] New endpoint to generate a client tenant token (#3531)
## Describe your changes

To be used for embedded components - the idea is that a tenant can generate a short-lived token on their backend that is bound to a single user, and specifically to a specified set of fields.

For example, if a tenant is displaying an embedded component for a card on their user's profile page, they would generate a short-lived token that has access to their authed user's `fp_id` and the credit card DI needed. 

The new decrypt API (in a follow-up PR) will only allow using this token to decrypt the provided piece of info

## Caveats

There's a lot more we can do with this - we could scope the token to only reads/writes for example. This might be more annoying to play with our auth model, but we can definitely add that functionality in the future

## Test Plan

Will test in future