Create VReqs inside the incode state in which they are needed (#3579)
## Describe your changes

In a future PR, we're going to start allowing the incode state machine to start up in states other than StartOnboarding and RetryUpload. The context associated with each state makes this really difficult - if we were starting up in, say, AddFront, we'd have to then find the corresponding VReq that was created in the last run. This just moves the VReq creation to the same time as the VRes

This breaks the initial intent of VReq - to (1) track outstanding requests that haven't been made yet and (2) to take a snapshot of the vault and be able to reconstruct the request we send to the vendor. The incode state machine now has its own metadata to track (1), and we compose the request sent to vendors from docv data and the incode state machine metadata

## Test Plan

ignored unit tests are currently broken while we do some incode surgery - will write more tests at the end of this migration :)