Create credentials when initializing incode state machine (#3581)
## Describe your changes

This does some fun surgery - the StartOnboarding state behaved so differently from all other states - it generates an auth token that all other states need in order to make requests to incode. So, it's more like initialization code than an independent state. This changes behavior to just instantly run the StartOnboarding state as soon as we initialize a machine. 

This seems a little unnecessary now, but it helps in a future PR: we're going to just pass the `session` in to each state's `run` function instead of making it state-specific context (since all states except StartOnboarding need it). This makes it much easier to start up the incode state machine in almost any state, which is necessary for progressive uploads