Make workflow actions generic/toplevel (#3588)
## Describe your changes

Right now Action's in Workflow land are scoped to a specific workflow.
This doesn't really make sense because a particular callsite shouldn't
need to know what kind of workflow it has in hand- it should just know
that a workflow exists and it is in a state that accepts the current
action at hand. For eg: in /authorize, we don't know if we have a
vanilla KYC workflow or a AlpacaKYC workflow, we just want to be able to
call `ww.run(Authorize)` and have it work for either case.

This PR makes `Action`'s a generic top level set of structs that
individual workflows share. Might make sense to do the same for State's
but that's a little more complex I think


## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->