Remove all last state-specific context (#3583)
## Describe your changes

One was user consent that can be fetched in the state.

Others were responses from previous states passed forward. Only real reason for this was to be able to extract information from the VReses on the `Complete` state, but application code doesn't do this anymore now that all business logic is in `Complete::enter`.

After this, the state machine will be able to bootstrap in ANY state!!!!