Fast Ephemeral Environments (#3676)
## Describe your changes
Brings down steady state first pulumi up from 20m -> < 6m (more can be done probably)

The major change is that use NeonDB instead of spinning up RDS which is what takes the bulk of the time. In the future we can use this to help test migration things too (Neon supports branching, which is what we use here, so we can branch from a dev-like DB with real fixture data).

I've added support for "Full Ephemeral" as well which brings up everything if you add that label (or if it detects changes in db.ts)

I also fixed a handful of other bring-up bring-down issues that caused failures previously on ephemeral envs.

![Screenshot 2023-05-29 at 11.50.41 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/uJtM6gX6vuLVCx8YhROH/0c67a2fe-639f-486b-8ae2-8d4f27cd7f7f/Screenshot%202023-05-29%20at%2011.50.41%20AM.png)